requirements.txt
asgiref==3.8.1
autopep8==2.3.2
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==24.1
django-js-asset==3.1.2
django-rest-multiple-models==2.1.3
django-rest-passwordreset==1.5.0
django-summernote==0.8.20.0
django-tinymce==4.1.0
django-unfold==0.49.1
djangorestframework==3.16.0
drf-spectacular==0.27.1
gunicorn==22.0.0
idna==3.10
joblib==1.4.2
packaging==24.2
pillow==11.1.0
pycodestyle==2.13.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==1.0.0
toml==0.10.2
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.9.0